+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Changelog
      `ResourceCollectionIterator`
    * moved all code to the `Puli\` top-level namespace
    * moved `ResourceRepository` and related classes to `Puli\Repository\`
+   * renamed `getByTag()` to `findByTag()`
 
 * 1.0.0-alpha3 (2014-02-22)